requirements.txt
click
python-gitlab
jinja2
requests
GitPython
dymport
